MAJOR: shctx/cache: Add way to have multiple shared_context/cache pairs
When the cache is used by too many threads at once (10+ threads), we end
up spending an enormous amount of time waiting for the shctx_lock to be
released since all threads compete on this single lock. This lock is
used to moved blocks from the avail list to the hot list and to reserve
blocks in which the responses will be cached.
Any lookup in the cache also requires this lock since when an entry is
found, its corresponding blocks are move to the hot list, which must be
done under lock.

The solution implemented in this patch consists in duplicating the
shared_context/cache pair into as many pairs are configured in the
cache. This can be set thanks to the "sub-caches" option. By default, we
will still have a single shared_context and cache instances per
configured cache. But if a specific number of sub-caches is defined, we
will allocate as many blocks as before in the shared_context block list
but those blocks won't be all added into the shared_context's avail list
but they will be split among all the shared_context lists. All of this
is done in the shctx_init function.

Whenever some cache related operation needs to be performed, we now need
to know on which sub-cache to operate. This is determined by the
request's (hash) mod (number of sub-caches) (see get_cache_from_hash
function).