MEDIUM: cache: Add multiple cache trees and locks
In order to limit contention on the cache tree we can use an array of
cache trees that will each have their own lock, indexed by the keys that
are already uint32.
The tree entries are aligned on 64 bits in order for each of them to be
on a separate cache line.
We also have one cleanup_list per cache which make the
cache_reserve_finish more costly because it must take as many locks are
the cache number in order to remove all the entries from their
corresponding cache.

The number of caches is hardcoded to 8 for now but it might be
configurable later.