MEDIUM: cache: Add refcount on cache_entry
Add a reference counter on the cache_entry. Its value will be atomically
increased and decreased via the retain_entry and release_entry
functions.
Do not take the lock in release_entry unless we are on the last
reference. This allows to limit the amount of times when we need to take
the cache lock during a release operation.