BUG/MINOR: ocsp: "Update successful" wrongly raised with unreachable ocsp responder
Since commit "88aa7a MINOR: http-client: Trigger an error if first
response block isn't a start-line", when an OCSP responder is
unreachable we exit the http client's io_handler earlier than before
without ever calling the stline callback.

This means that in case of unreachable OCSP responder, we ended up
reaching the end of the response (as-in calling the response_end
callback) without ever calling any ever httpclient callback. We ended up
wrongly thinking that the update was successful.

Now, instead of catching the error thanks to the 'hc->res.status != 200'
test we had during stline processing, we need to manage the case where
the end of a response is reached but no body was ever found.

This patch does not need to be backported.