MINOR: cache: Rename 'release_entry_XXX' functions for better consistency
In order to match the logic of other functions suffixed _locked or
_unlocked (such as __ssl_ocsp_update_insert_unlocked or
__ckch_inst_free_locked for instance), the release_entry_locked function
will now take the cache lock all the time and the regular release_entry
function will not take the lock and expect that the caller already has
the proper lock taken.
Basically, all the previous calls to 'release_entry_locked' or to the
previous 'release_entry' function with a 'needs_locking==0' parameter
are now replaced by a simple 'release_entry' call that does not take the
lock. All the previous calls to 'release_entry_unlocked' or to
'release_entry' with 'needs_locking==1' will now be replaced by a call
to 'release_entry_locked'.